chore: refactor executors cache out from bank.rs
